Move diagnostic files into its own folder
