+++ src/components/diagnostics/diagnostics-pane.tsx
